requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
bleach==6.2.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
cron_descriptor==2.0.5
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
Django==5.2.5
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.1
evaluate==0.4.5
executing==2.2.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kombu==5.5.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.1
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1
python-decouple==3.8
django-cors-headers==4.3.1
